requirements.txt
fastapi
uvicorn
pillow
torch
torchvision
